Phpstan level 7 (#1380)

+++ src/Tracing/Tracing.php
@@ -130,7 +130,7 @@ class Tracing
     protected function diffTimeInNanoseconds($start, $end): int
     {
         if ($this->platformSupportsNanoseconds()) {
-            return $end - $start;
+            return (int) ($end - $start);
         }
 
         // Difference is in seconds (with microsecond precision)
